requirements.txt
pillow~=9.0.0
keras~=2.7.0
tensorflow~=2.7.0
tensorflow-gpu~=2.7.0
numpy~=1.21.5
python-dotenv~=0.20.0
flask~=2.0.3
